Bug 568393 - msvc 2008 compiler generates bad code for CallInfo static initializer (r+wmaddox)
ugh

--HG--
extra : convert_revision : f7c07b8ca96fc2ef1179171f05e69b841ac2865f

+++ js/src/nanojit/LIR.h
@@ -323,7 +323,7 @@ namespace nanojit
         uintptr_t   _address;
         uint32_t    _typesig:27;     // 9 3-bit fields indicating arg type, by ARGTYPE above (including ret type): a1 a2 a3 a4 a5 ret
         AbiKind     _abi:3;
-        uint8_t     _isPure:1;      // _isPure=1 means no side-effects, result only depends on args
+        uint32_t    _isPure:1;      // _isPure=1 means no side-effects, result only depends on args
         AccSet      _storeAccSet;   // access regions stored by the function
         verbose_only ( const char* _name; )
